chore: add ToastMark, Renderer interface
