+++ libs/toastmark/src/commonmark/custom/__test__/customInline.spec.ts
@@ -1,5 +1,5 @@
 import { Parser } from '../../blocks';
-import { Renderer } from '../../../html/render';
+import { Renderer } from '../../../html/renderer';
 import { CustomInlineNode } from '../../node';
 
 const reader = new Parser();
