+++ libs/toastmark/src/index.ts
@@ -1,3 +1,3 @@
 export { ToastMark } from './toastmark';
-export { Renderer } from './html/render';
+export { Renderer } from './html/renderer';
 export { Parser } from './commonmark/blocks';
