+++ libs/toastmark/types/index.d.ts
@@ -1,234 +1,3 @@
-export type BlockNodeType =
-  | 'document'
-  | 'list'
-  | 'blockQuote'
-  | 'item'
-  | 'heading'
-  | 'thematicBreak'
-  | 'paragraph'
-  | 'codeBlock'
-  | 'htmlBlock'
-  | 'table'
-  | 'tableHead'
-  | 'tableBody'
-  | 'tableRow'
-  | 'tableCell'
-  | 'tableDelimRow'
-  | 'tableDelimCell'
-  | 'refDef'
-  | 'customBlock'
-  | 'frontMatter';
-
-export type InlineNodeType =
-  | 'code'
-  | 'text'
-  | 'emph'
-  | 'strong'
-  | 'strike'
-  | 'link'
-  | 'image'
-  | 'htmlInline'
-  | 'linebreak'
-  | 'softbreak'
-  | 'customInline';
-
-export type MdNodeType = BlockNodeType | InlineNodeType;
-
-export type SourcePos = [[number, number], [number, number]];
-
-export interface NodeWalker {
-  current: MdNode | null;
-  root: MdNode;
-  entering: boolean;
-
-  next(): { entering: boolean; node: MdNode } | null;
-  resumeAt(node: MdNode, entering: boolean): void;
-}
-
-export interface MdNode {
-  type: MdNodeType;
-  id: number;
-  parent: MdNode | null;
-  prev: MdNode | null;
-  next: MdNode | null;
-  sourcepos?: SourcePos;
-  firstChild: MdNode | null;
-  lastChild: MdNode | null;
-  literal: string | null;
-
-  isContainer(): boolean;
-  unlink(): void;
-  replaceWith(node: MdNode): void;
-  insertAfter(node: MdNode): void;
-  insertBefore(node: MdNode): void;
-  appendChild(child: MdNode): void;
-  prependChild(child: MdNode): void;
-  walker(): NodeWalker;
-}
-
-export interface BlockMdNode extends MdNode {
-  type: BlockNodeType;
-
-  // temporal data (for parsing)
-  open: boolean;
-  lineOffsets: number[] | null;
-  stringContent: string | null;
-  lastLineBlank: boolean;
-  lastLineChecked: boolean;
-}
-
-export interface ListData {
-  type: 'ordered' | 'bullet';
-  tight: boolean;
-  start: number;
-  bulletChar: string;
-  delimiter: string;
-  markerOffset: number;
-  padding: number;
-  task: boolean;
-  checked: boolean;
-}
-
-export interface ListMdNode extends BlockMdNode {
-  listData: ListData | null;
-}
-
-export interface ListItemMdNode extends BlockMdNode {
-  parent: NonNullable<MdNode>;
-  listData: ListData | null;
-}
-
-export interface HeadingMdNode extends BlockMdNode {
-  level: number;
-  headingType: 'atx' | 'setext';
-}
-
-export interface CodeBlockMdNode extends BlockMdNode {
-  fenceOffset: number;
-  fenceLength: number;
-  fenceChar: string | null;
-  info: string | null;
-  infoPadding: number;
-}
-
-export interface TableColumn {
-  align: 'left' | 'center' | 'right';
-}
-
-export interface TableMdNode extends BlockMdNode {
-  columns: TableColumn[];
-}
-
-export interface TableCellMdNode extends BlockMdNode {
-  startIdx: number;
-  endIdx: number;
-  paddingLeft: number;
-  paddingRight: number;
-  ignored: boolean;
-}
-
-export interface RefDefMdNode extends BlockMdNode {
-  title: string;
-  dest: string;
-  label: string;
-}
-
-export interface CustomBlockMdNode extends BlockMdNode {
-  syntaxLength: number;
-  offset: number;
-  info: string;
-}
-
-export interface HtmlBlockMdNode extends BlockMdNode {
-  htmlBlockType: number;
-}
-
-export interface LinkMdNode extends MdNode {
-  destination: string | null;
-  title: string | null;
-  extendedAutolink: boolean;
-}
-
-export interface CodeMdNode extends MdNode {
-  tickCount: number;
-}
-
-export interface ImageMdNode extends MdNode {
-  lastChild: MdNode;
-  destination: string;
-  title: string;
-}
-
-export interface CustomInlineMdNode extends MdNode {
-  info: string;
-}
-
-/* ToastMark Parser type */
-export type CustomParser = (node: MdNode, context: { entering: boolean }) => void;
-export type CustomParserMap = Partial<Record<MdNodeType, CustomParser>>;
-
-export type RemovedNodeRange = {
-  id: [number, number];
-  line: [number, number];
-};
-
-export interface EditResult {
-  nodes: MdNode[];
-  removedNodeRange: RemovedNodeRange | null;
-}
-
-/* ToastMark Renderer type */
-export interface TagToken {
-  tagName: string;
-  outerNewLine?: boolean;
-  innerNewLine?: boolean;
-}
-
-export interface OpenTagToken extends TagToken {
-  type: 'openTag';
-  classNames?: string[];
-  attributes?: Record<string, string>;
-  selfClose?: boolean;
-}
-
-export interface CloseTagToken extends TagToken {
-  type: 'closeTag';
-}
-
-export interface TextToken {
-  type: 'text';
-  content: string;
-}
-
-export interface RawHTMLToken {
-  type: 'html';
-  content: string;
-  outerNewLine?: boolean;
-}
-
-export type HTMLToken = OpenTagToken | CloseTagToken | TextToken | RawHTMLToken;
-
-export type CustomHTMLRenderer = (
-  node: MdNode,
-  context: Context,
-  convertors?: CustomHTMLRendererMap
-) => HTMLToken | HTMLToken[] | null;
-
-export type CustomHTMLRendererMap = Partial<Record<string, CustomHTMLRenderer>>;
-
-export interface ContextOptions {
-  gfm: boolean;
-  softbreak: string;
-  nodeId: boolean;
-  tagFilter: boolean;
-  convertors?: CustomHTMLRendererMap;
-}
-
-export interface Context {
-  entering: boolean;
-  leaf: boolean;
-  options: Omit<ContextOptions, 'convertors'>;
-  getChildrenText: (node: MdNode) => string;
-  skipChildren: () => void;
-  origin?: () => ReturnType<CustomHTMLRenderer>;
-}
+export * from './node';
+export * from './toastMark';
+export * from './renderer';
